pyproject.toml
[project]
name = "pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nicolas Brosse", email = "nbrosse@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.1",
    "google-genai>=1.2.0",
    "jinja2>=3.1.5",
    "llama-index-core>=0.12.19",
    "llama-index-embeddings-gemini>=0.3.1",
    "llama-index-llms-gemini>=0.4.10",
    "llama-index-readers-file>=0.4.5",
    "llama-index-retrievers-bm25>=0.5.2",
    "mistralai>=1.5.0",
    "pdf2image>=1.17.0",
    "pypdf>=5.3.0",
    "python-dotenv>=1.0.1",
    "shiny>=1.3.0",
    "tenacity>=9.0.0",
]

[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "ipython>=8.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
